Absorb TreeGrid into WaydGrid2 as tree mode
WaydGrid2 is now the one grid: flat by default, and providing getSubRows
turns on tree mode - row expansion with filterFromLeafRows (a matching
child keeps its ancestor chain visible), the tree row form behind the
row-renderer seam (sortable-row wrapper, per-cell data-cell-id, editing
affordances, validation error rows), reparenting drag-and-drop, inline
editing, and draft rows. The editing, draft, and DnD machinery moved in
from TreeGrid unchanged; tree-utils and draft-utils moved to
wayd-grid-core with their tests.

Migrate the three tree screens (roadmap items, project plan, project
work items) onto WaydGrid2 tree mode. Their inline filter row is
replaced by the descriptor filter system (floating filters + filter
popups): explicit filterFn props are dropped in favor of the engine
defaults and meta.filterType, which upgrades text-based date/number
filters to the typed filter panels. The screens also adopt the shared
toolbar and the unified grid styling; indent/caret/drag-handle cell
classes moved into the grid CSS module. Delete tree-grid entirely.

Also fix the loading/empty status band: center the spinner and empty
message in the grid body (flex on an inner div, not the td, and the
table stretches to fill the wrapper while the band shows) and use the
large spinner to match the ag-grid WaydGrid overlay.
