--- Wayd.Web/src/wayd.web.reactclient/src/components/common/tree-grid/tree-grid.module.css
@@ -1,378 +1,0 @@
-.table {
-  width: 100%;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  min-height: 0;
-}
-
-.toolbar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: var(--ant-margin);
-  margin-bottom: var(--ant-margin);
-  flex-wrap: nowrap;
-  overflow: hidden;
-}
-
-.toolbarRight {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  gap: var(--ant-margin-sm);
-  flex-wrap: nowrap;
-  white-space: nowrap;
-}
-
-.toolbarSearch {
-  width: 225px;
-}
-
-.toolbarDivider {
-  width: 1px;
-  height: 30px;
-  background-color: var(--ant-color-split);
-}
-
-.tableWrapper {
-  overflow-x: auto;
-  overflow-y: auto;
-  border-radius: var(--ant-border-radius-lg);
-  border: 1px solid var(--ant-color-border);
-  background-color: var(--ant-color-bg-container);
-  flex: 1;
-  min-height: 0;
-}
-
-.tableElement {
-  --tree-grid-header-row-height: 29px;
-  width: max-content;
-  min-width: 100%;
-  border-collapse: separate;
-  border-spacing: 0;
-  background-color: transparent;
-  font-size: var(--ant-font-size);
-  table-layout: fixed;
-}
-
-
-.th {
-  padding: 5px var(--ant-padding-sm);
-  background-color: var(--ant-color-bg-container);
-  border-bottom: 1px solid var(--ant-color-border);
-  color: var(--ant-color-text);
-  text-align: left;
-  font-size: 12px;
-  font-weight: var(--ant-font-weight-strong);
-  white-space: nowrap;
-  position: sticky;
-  top: 0;
-  z-index: 20;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 0;
-}
-
-.thSortable {
-  cursor: pointer;
-  -webkit-user-select: none;
-  user-select: none;
-}
-
-.thResizable {
-  padding-left: 0;
-}
-
-.thResizable .thContent {
-  padding-right: 12px;
-}
-
-.thContent {
-  display: inline-flex;
-  align-items: center;
-  gap: var(--ant-margin-xs);
-  max-width: 100%;
-  min-width: 0;
-  overflow: hidden;
-}
-
-.thText {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  min-width: 0;
-  flex: 1;
-}
-
-.resizer {
-  position: absolute;
-  top: 0;
-  right: 0;
-  width: 10px;
-  height: 100%;
-  cursor: col-resize;
-  -webkit-user-select: none;
-  user-select: none;
-  touch-action: none;
-  z-index: 1;
-}
-
-.resizer::after {
-  content: '';
-  position: absolute;
-  top: var(--ant-padding-xs);
-  bottom: var(--ant-padding-xs);
-  right: 4px;
-  width: 1px;
-  background-color: var(--ant-color-border);
-  opacity: 0.7;
-}
-
-.resizerActive::after {
-  background-color: var(--ant-color-primary);
-  opacity: 1;
-}
-
-.filterTh {
-  padding: 4px 4px;
-  background-color: var(--ant-color-bg-container);
-  border-bottom: 1px solid var(--ant-color-border);
-  color: var(--ant-color-text);
-  font-weight: normal;
-  overflow: hidden;
-  text-align: left;
-  position: sticky;
-  top: var(--tree-grid-header-row-height);
-  z-index: 19;
-}
-
-.filterControl {
-  width: 100%;
-  max-width: 100%;
-  box-sizing: border-box;
-}
-
-.filterControl.antInputAffixWrapper {
-  min-width: 0;
-  max-width: 100%;
-}
-
-.filterControl.antInputAffixWrapper :global(.ant-input) {
-  min-width: 0;
-  font-weight: normal;
-}
-
-.filterControl.antSelect {
-  width: 100%;
-  min-width: 0;
-  max-width: 100%;
-  position: relative;
-}
-
-.filterControl.antSelect :global(.ant-select-selector) {
-  min-width: 0;
-  max-width: 100%;
-  align-items: center;
-  flex-wrap: nowrap !important;
-  overflow: hidden;
-}
-
-.filterControl.antSelect :global(.ant-select-selection-item) {
-  font-weight: normal;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  flex-shrink: 1;
-}
-
-.filterControl.antSelect :global(.ant-select-selection-overflow) {
-  flex-wrap: nowrap !important;
-  overflow: hidden;
-}
-
-.filterControl.antSelect :global(.ant-select-selection-overflow-item) {
-  flex-shrink: 1;
-  max-width: 100%;
-}
-
-.filterControl.antSelect :global(.ant-select-selection-search) {
-  min-width: 0;
-}
-
-.filterControl.antSelect :global(.ant-select-selection-search-input) {
-  font-weight: normal;
-}
-
-.filterControl.antSelect :global(.ant-select-clear) {
-  margin-right: 10px;
-}
-
-.filterPopup {
-  min-width: 220px;
-}
-
-.td {
-  padding: 2px 7px;
-  border-bottom: 1px solid var(--ant-color-border-secondary);
-  color: var(--ant-color-text);
-  font-size: 12px;
-  vertical-align: middle;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  max-width: 0;
-}
-
-.tr {
-  cursor: default;
-}
-
-.trEditable {
-  cursor: pointer;
-}
-
-.tr:hover .td {
-  background-color: var(--ant-color-fill-secondary);
-}
-
-.trAlt .td {
-  background-color: var(--ant-color-fill-quaternary);
-}
-
-.trSelected .td {
-  background-color: var(--ant-color-bg-container);
-}
-
-.editableCell {
-  padding: 1px 4px !important;
-}
-
-.editableCell :global(.ant-input),
-.editableCell :global(.ant-select),
-.editableCell :global(.ant-picker) {
-  width: 100%;
-  font-size: 12px;
-}
-
-.expanderBtn {
-  padding: 0 4px;
-}
-
-.expanderBtn :global(.anticon) {
-  font-size: 12px;
-}
-
-.nameCell {
-  display: flex;
-  align-items: center;
-  gap: var(--ant-margin-xxs);
-  min-width: 0;
-  width: 100%;
-}
-
-.indentSpacer {
-  display: inline-block;
-  width: 24px;
-  flex: 0 0 24px;
-}
-
-.phaseName {
-  font-weight: 600;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.empty {
-  padding: var(--ant-padding-lg);
-}
-
-.loading {
-  padding: var(--ant-padding-lg);
-  display: flex;
-  justify-content: center;
-}
-
-.validationErrorRow {
-  background-color: var(--ant-color-error-bg);
-}
-
-.validationErrorCell {
-  color: var(--ant-color-error);
-  font-size: 12px;
-  padding: 12px;
-  border-top: 1px solid var(--ant-color-error-border);
-}
-
-.validationErrorItem {
-  margin-bottom: 4px;
-}
-
-.validationErrorItem:last-child {
-  margin-bottom: 0;
-}
-
-.validationErrorField {
-  font-weight: 500;
-}
-
-/* Drag and Drop Styles */
-.dragHandle {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 4px;
-  transition: opacity 0.2s;
-}
-
-.dragHandle:hover {
-  opacity: 0.8;
-}
-
-.inlineRowAction {
-  opacity: 0;
-  pointer-events: none;
-  transition: opacity 0.15s ease;
-}
-
-.rowActionBtn {
-  padding: 0 4px;
-  height: 20px;
-  font-size: 12px;
-  flex-shrink: 0;
-}
-
-.colorPickerFocusTarget {
-  display: inline-flex;
-  width: 100%;
-  outline: none;
-}
-
-.colorPickerFocusTarget:focus {
-  outline: none;
-}
-
-.colorPickerFocusTarget:focus-visible :global(.ant-color-picker-trigger),
-.colorPickerFocusTarget:focus-within :global(.ant-color-picker-trigger) {
-  border-color: var(--ant-color-primary);
-  box-shadow:
-    0 0 0 var(--ant-control-outline-width, 2px) var(--ant-control-outline);
-  outline: none;
-}
-
-.tr:hover .inlineRowAction,
-.trSelected .inlineRowAction {
-  opacity: 1;
-  pointer-events: auto;
-}
-
-.tr[data-dragging='true'] {
-  opacity: 0.4;
-  position: relative;
-  z-index: 999;
-}
-
-.tr[data-dragging='true'] .td {
-  background-color: var(--ant-color-fill-tertiary);
-}
